A few minor build fixes (#230)
1. Set SPARK_LOCAL_IP and increase spark.port.maxRetries to hopefully fix test flakiness for Spark
2. Automatically delete the rikai ivy cache when publishLocal.

publishLocal sends the jar to ivy local repository. When we run tests which refers to
the rikai jar as a spark package, it loads the local jar and caches it. This then makes it
so that subsequent publishLocal is shadowed by the cached jar unless it is manually removed.
Instead we override publishLocal task to first delete the ivyHome/cache/ai.eto/rikai_2.12 directory